+++ R/docs_types.R
@@ -18,7 +18,7 @@
 #' Return types describe the `zod` schema of objects that Rserve functions,
 #' and most of these utilise the `Robj` utility types in the `rserve-ts`
 #' library. The return types have additional properties added,
-#' namedly `r_type` and `r_attributes`, handled by the `Robj` utility types.
+#' namely `r_type` and `r_attributes`, handled by the `Robj` utility types.
 #'
 #' **Scalar versus array ("vector") types**:
 #' In R, almost all types are vectors. In the 'rserve-js' library,
@@ -27,7 +27,7 @@
 #' has unknown length. `which(x > 5)` is one such example.
 #'
 #' To solve this, we add an `n` argument to the `ts_*` functions. When `n = 1`,
-#' the type takes the *scalar* form of the alue. When `n != 1`, the type takes
+#' the type takes the *scalar* form of the value. When `n != 1`, the type takes
 #' the *array* form of the value (this includes 0). Otherwise, the type
 #' is the union of the scalar and array forms.
 #'
@@ -35,13 +35,13 @@
 #'
 #' # Available types
 #'
-#' - `ts_boolean`: A boolean value. The array type is `Int8Array`.
+#' - `ts_logical`: A logical value. The array type is `Int8Array`.
 #' - `ts_integer`: An integer value. The array type is `Int32Array`.
 #'                 Javascript does not have a native integer type,
 #'                 so scalars are represented as a number
 #'                 (the same as `ts_numeric`).
 #' - `ts_numeric`: A numeric value. The array type is `Float64Array`.*
-#' - `ts_string`: A string value. The array type is `string[]`.
+#' - `ts_character`: A character value. The array type is `string[]`.
 #' - `ts_factor`: A factor value. The array type is `(level1 | level2 | ... | levelN)[]`, and this type does not have a scalar form.
 #' - `ts_list`: A list value, represented by a named object or an array.
 #' - `ts_dataframe`: A data frame value, represented by a named object.