+++ tests/testthat/test-compile.R
@@ -5,7 +5,10 @@ test_that("anonomous functions", {
     )
 
     add_c <- ts_compile(add)
-    expect_equal(add_c, "export const add = Robj.ocap([z.number(), z.number()], Robj.numeric(1));")
+    expect_equal(
+        as.character(add_c),
+        "export const add = Robj.ocap([z.number(), z.number()], Robj.numeric(1));"
+    )
 })